+++ core/src/com/angian/bobby/Sausage.java
@@ -4,7 +4,7 @@ import com.badlogic.gdx.math.Rectangle;
 import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.scenes.scene2d.Stage;
 
-import static com.angian.bobby.LevelConstants.FRAME_DURATION;
+import static com.angian.bobby.LevelConstants.ANIM_FRAME_DURATION;
 
 
 public class Sausage extends BaseActor {
@@ -12,7 +12,7 @@ public class Sausage extends BaseActor {
 
     public Sausage(LevelScreen.EnemyHeight height, Stage s) {
         super(s);
-        loadAnimationFromSheet("sausage_spritesheet.png", 2, 1, FRAME_DURATION, true);
+        loadAnimationFromSheet("sausage_spritesheet.png", 2, 1, ANIM_FRAME_DURATION, true);
 
 
         if (height == LevelScreen.EnemyHeight.HIGH)